Create `plan` module in `ui`
