+++ src/ui.rs
@@ -1,8 +1,10 @@
 pub mod markdown;
+pub mod plan;
 pub mod reply;
 
 mod reasoning;
 
 pub use markdown::Markdown;
+pub use plan::Plan;
 pub use reasoning::Reasoning;
 pub use reply::Reply;
